Refactor code to separate typeDefs and resolvers
The typeDefs and resolvers code has been separated from the index file and moved to their respective new files in the graphql/schema folder. This provides better organization of the code and makes it easier to manage and update in the future. Additionally, the yoga function call in the index file has been updated to include debug logging.