Add URL checker utility in backend
The new JavaScript file 'url_checker.js' has been added in the backend codebase. It includes the function that performs asynchronous checks on a list of URLs to determine their validity. This list of URLs corresponds to expected image locations. Reports are generated revealing the count of valid and invalid URLs.