Refactor product view and update backend dependencies
The code for the product view page and image error handling was refactored for better readability and efficiency. Small changes to the CSS presentation of the product grid were also made. In the backend, override rules for GraphQL packages were added, along with new trusted dependencies in 'package.json'. A couple of constants related to the product view were moved for better placement in '+page.server.ts'.