"Implement new collection features and cleanup redundant code"
A new feature has been implemented to provide more details in the collection view. The code now includes a filter and search system to narrow down the products being displayed. In the process of doing this, some redundant and commented-out code was removed, including unnecessary test pages and configurations. The changes also involved some adjustments to the layout, GraphQL queries and generating breadcrumbs.