Implement page view tracking with OpenReplay
This commit introduces the OpenReplay event tracking for page views across various pages in the application. Tracking is implemented using the `getTracker` function, invoked on each page's `onMount` lifecycle method. Additionally, a global initialization and context setup for the OpenReplay tracker is done in the base layout component. This also includes creation of a translation utility for translating codes into readable strings to be used in tracking data.