Add look details query and display related styles
This commit incorporates a new look details query in the GraphQL resolvers. It fetches and displays related styles of a selected item in the frontend. Alongside, it also adds a new "LookDetails" type in GraphQL typedefs and frontend types. This new feature refines the item details view providing additional relevant information to the user.