Update failed_urls and url_suffixes content
The failed_urls.txt file and the url_suffixes.json file were updated. Several URLs were removed and new ones added in the failed_urls.txt. Also, there was a significant reduction in the number of URLs in the url_suffixes.json file. These changes are likely connected with changes in the website structure or data that is being tracked.